Refactor program to be more Object-Oriented
This change adds methods to Player, Food, and Team such as Player#canSee(Food)
and Food#seenBy(), so that methods in Util class can be removed.

With this change, a new class MessagePlayerId is added in place of Player,
sent on client connection, because there seems to be an issue where
netty-socketio tries to send all of the instance variables including the
BarngaOnlineConfigsDefault instance variable. MessagePlayerId is just a
container class with only the necessary information when sending the player's
identity (basically same as Player instance variables without the
BarngaOnlineConfigsDefault).